2025-07-10(목) sign

+++ src/main/java/kr/co/bestiansoft/ebillservicekg/sed_jk/dto/request/SignRequestDto.java
@@ -3,7 +3,7 @@ package kr.co.bestiansoft.ebillservicekg.sed_jk.dto.request;
 import jakarta.validation.constraints.NotNull;
 
 public record SignRequestDto(
-        Long apvlId,
+        int apvlId,
         String pinCode
 ) {
 }
